requirements.txt
# Kafka
kafka-python

# Apache Airflow
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-ssh
apache-airflow-providers-http

# Apache Spark
pyspark

# Apache Cassandra
cassandra-driver

# HTTP Requests
requests
